학원 테이블 리뷰 카운트 칼럼 추가 건

+++ GC25 Project/src/main/java/com/gc25/dto/AcademyDTO.java
@@ -7,7 +7,7 @@ public class AcademyDTO {
 	String academyTel; //학원 전화번호
 	String academyAddress; //학원 주소
 	String academyRodeAddress; //학원 도로주소
-	int academyAvgScore; //학원 만족도
+	double academyAvgScore; //학원 만족도
 	String academyUrl; 
 	String academyX; //위치 경도 
 	String academyY; //위치 위도
@@ -19,7 +19,7 @@ public class AcademyDTO {
 
 
 	public AcademyDTO(int academyNumber, String academyName, String academyTel, String academyAddress,
-			String academyRodeAddress, int academyAvgScore, String academyUrl, String academyX, String academyY) {
+			String academyRodeAddress, double academyAvgScore, String academyUrl, String academyX, String academyY) {
 		super();
 		this.academyNumber = academyNumber;
 		this.academyName = academyName;
@@ -83,12 +83,12 @@ public class AcademyDTO {
 	}
 
 
-	public int getAcademyAvgScore() {
+	public double getAcademyAvgScore() {
 		return academyAvgScore;
 	}
 
 
-	public void setAcademyAvgScore(int academyAvgScore) {
+	public void setAcademyAvgScore(double academyAvgScore) {
 		this.academyAvgScore = academyAvgScore;
 	}
